@@ -4,7 +4,7 @@
  */
 package uk.co.philpotter.examcalc;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.View;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
