@@ -6,7 +6,7 @@ package uk.co.philpotter.examcalc;
 
 import android.app.AlertDialog;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
